requirements.txt
appdirs==1.4.1
certifi==2017.4.17
chardet==3.0.3
click==6.7
decorator==4.0.11
dj-database-url==0.4.2
Django==1.10.5
djangorestframework==3.5.4
geocoder==1.22.6
gunicorn==19.6.0
idna==2.5
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
ratelim==0.1.6
requests==2.17.3
six==1.10.0
urllib3==1.21.1
whitenoise==3.3.0
